ModelCheckPoint added and changed the loss plot
The ModelcChackPoint monitors the validation loss not the validation accuracy. It saves the model in a path ( /content/best_model.h5' of colab in this case . Change it with your path) and after fitting the model it loads the best model in autoencoder. -->  self.autoencoder=load_model('/content/best_model.h5')